Extract shared JSONL usage-entry parser
Move the parse + dedup-key logic out of UsageMonitor.processLine into
lib/entry-parser.js so the live monitor and the new history scanner share
one implementation. The monitor keeps its own today-window filter and
dedup set; only the record parsing is shared.
