+++ lib/usage-monitor.js
@@ -3,8 +3,8 @@
 const fs = require('fs');
 const path = require('path');
 const os = require('os');
-const crypto = require('crypto');
 const { costForEntry } = require('./pricing');
+const { parseUsageEntry } = require('./entry-parser');
 
 /**
  * Monitors the *.jsonl session logs under ~/.claude/projects for assistant usage entries,
@@ -170,22 +170,11 @@ class UsageMonitor {
   }
 
   processLine(lineBuffer) {
-    let obj;
-    try {
-      obj = JSON.parse(lineBuffer.toString('utf8'));
-    } catch {
-      return;
-    }
-    if (!obj || obj.type !== 'assistant') return;
-    if (typeof obj.timestamp !== 'string') return;
-    const message = obj.message;
-    if (!message || typeof message !== 'object') return;
-    const usageDict = message.usage;
-    if (!usageDict || typeof usageDict !== 'object') return;
+    const parsed = parseUsageEntry(lineBuffer);
+    if (!parsed) return;
 
     // Local-day filter.
-    const timestamp = new Date(obj.timestamp);
-    if (isNaN(timestamp.getTime())) return;
+    const timestamp = parsed.timestamp;
     const now = new Date();
     if (
       timestamp.getFullYear() !== now.getFullYear() ||
@@ -197,40 +186,12 @@ class UsageMonitor {
     if (timestamp.getTime() < this.currentDayStart.getTime()) return;
 
     // Dedup key.
-    const requestId = typeof obj.requestId === 'string' ? obj.requestId : '';
-    const msgId = typeof message.id === 'string' ? message.id : crypto.randomUUID();
-    const key = requestId + ':' + msgId;
-    if (this.seenKeys.has(key)) return;
-    this.seenKeys.add(key);
+    if (this.seenKeys.has(parsed.key)) return;
+    this.seenKeys.add(parsed.key);
 
-    const intValue = (v) => (typeof v === 'number' && isFinite(v) ? Math.trunc(v) : 0);
-
-    const inputTokens = intValue(usageDict.input_tokens);
-    const outputTokens = intValue(usageDict.output_tokens);
-    const cacheReadInputTokens = intValue(usageDict.cache_read_input_tokens);
-    const cacheCreationInputTokens = intValue(usageDict.cache_creation_input_tokens);
-
-    let cacheCreationBreakdown = null;
-    const cc = usageDict.cache_creation;
-    if (cc && typeof cc === 'object') {
-      cacheCreationBreakdown = {
-        ephemeral5mInputTokens: intValue(cc.ephemeral_5m_input_tokens),
-        ephemeral1hInputTokens: intValue(cc.ephemeral_1h_input_tokens),
-      };
-    }
-
-    const entry = {
-      inputTokens,
-      outputTokens,
-      cacheReadInputTokens,
-      cacheCreationInputTokens,
-      cacheCreationBreakdown,
-    };
-
-    const model = typeof message.model === 'string' ? message.model : null;
-    this.snapshot.totalCostUSD += costForEntry(entry, model);
-    this.snapshot.inputTokens += inputTokens;
-    this.snapshot.outputTokens += outputTokens;
+    this.snapshot.totalCostUSD += costForEntry(parsed.usage, parsed.model);
+    this.snapshot.inputTokens += parsed.inputTokens;
+    this.snapshot.outputTokens += parsed.outputTokens;
     this.snapshot.entryCount += 1;
   }
 }
